Fix error message when entering crypto amount
Simplex API is returning a glitchy and incorrect error string when a user enters a crypto amount below the minimum threshold. The minimum amount keeps increasing as the user types in new values. Detect this error string and simply return the limits in a fixed amount of USD.

This is admittedly a temporary fix until we fully rewrite the plugin in native code